Uploading an unfinished transliterator to Bengali.
I have been working on this for a while now. It is very far from complete. Maybe it won't fit inside one script. Only time will tell.
The aim is to produce a POSIX-compatible transliterator to the Bengali language. It reads ASCII and produces Bengali characters in the output. I/O is posix compatible.