Added a new file for viewing CSV in pretty tables inside the terminal
`csview` is a Python3 based utility which opens CSV files inside the terminal itself. It allows you to view the table in a proper tabulation, unlike less or cat which gives an ugly unformatted text. Simple place this file inside your path, and check to make sure that the Python3 application is where the shebang says it is.
In case you have any issues or questions, please feel free to ask me.